Restructured caching docs
Don't throw so much detail at the reader early on in the doc,
this is supposed to be an overview of the different caching approaches.
In particular, we should highlight the simplicity and benefits of HTTP based caching,
so I've pulled that up on the page.

I've also split up the "how to identify" examples into distinct practice sections,
to make it flow a bit nicer.