Exclude Files from search query when ShowInSearch is false
At present, the `CwpSearchEngine` will exclude `SiteTree` items from search when "ShowInSearch" is false.
`File` has a "ShowInSearch" boolean field as well, but it is not (by default) editable via the CMS.
However, if someone chooses to exposes this in the CMS (because they may want their Admins to be able to set certain files to not be searchable) and they are using the default `CwpSearchEngine`, they will still get all Files returned / indexed.

This change adds an exclude to the default search query in `CwpSearchEngine` to not show any `File` that has "ShowInSearch" set to false.