FIX swap GridField re-ordering component
`undefinedoffset/sortablegridfield` is a good tool, however will not
be supported in the future by our (SilverStripe Ltd.) maintenance team.
The other option of `symbiote/silverstripe-gridfieldextensions` however
will remain supported (as it includes more functionality than just
sorting rows), so we'll use this and add it as a dependency here
instead. We are doing this to reduce maintenance overhead of two modules
that solve the same problem.