FIX Unentice direct BasePage creation in the CMS
The extension for TaxonomyTerm applies a relation to Page,
which is BasePage. The nature of BasePage is that is _should_ be
`abstract`, and should not be created. Unfortunately for us the
default scaffolder offers a CMS user the option to create a base page
via the relation to TaxonomyTerm. This commit stops that.