Unbreak Backwards Compatibility for RelatedPages
With the recent update to use ManyManyThroughList instead of the old
ManyManyList for the 'RelatedPages' feature on BasePage, this
has altered the return type of the applicable getters, and the two types
of many many relationship do unfortuantely not implement the same
interface. This breaks backwards compatiblity, so this commit in turn
attempts to preserve it through keeping both types of relationship.

Using the through relationship to maintain the ability to order ones
preferences on the related pages list, while keeping the default
'RelatedPages' relationship to return the ManyManyList where anyone
might be expecting it.

In the background they use the same relationship information, most
applicably the same join table that actually defines the relationships
through the new SilverStripe 4 feature to be able to custom name a
model's database table. This has worked in our advantage by naming the
same table as is generated in the old style many_many relationship.