Update gridfieldextensions to handle manymanyThrough
The alterations to support ordering many many through lists in
gridfieldextensions will be released as 3.2.0 - the current requirements
of cwp/cwp (this module) require ^3.1 - which will pull in the version
we need, however it's better to update the minimum version since the
update to features here rely upon it. It is improbable yet plausible
that someone could end up with an inability to sort their many_many
releated pages releationship, which is what parent commits of this one
are attempting to prevent.