DOCS add post-2020 note to changlog index
The CWP Lead Agency agreement ends in September 2020. As some CWP versions will be supported for longer than the Lead Agency Agreement, we're adding a note to a 'service continuation' page on cwp.govt.nz to provide further information.

This will also need to be merged up to master to be reflected on cwp.govt.nz.

I'm not entirely sure on whether this sentence needs to be on the same line as the `introduction:` tag though...