Update the 1.9.3 release docs re: MFA again
MFA 3.0.0 release included a conflicts directive with Subsites 1.4.1,
where the previous advice in the release docs was to alias 1.4.2 as 1.4.1
this saw the dependency resolve, but then also triggered the confliction.

The best way around this was decided to provide a more detailed alteration
to the project composer.json - as this will only last for 2 months (next
CWP relase will resolve this need), and any other constraint relaxation
would be unfair to other consumers of the MFA product with subsites.