Remove docs document that should be independent of the recipe.
Platform changes independently of the recipe, so there is no point in
tying this document with the recipe - it becomes outdated very quickly,
and there is no easy way to fix it for past tagged releases.

This commit removes the the "Infrastructural considerations" document
from here. From now on it will be added during the docs update process
inside the CWP portal, by being downloaded from the internal
platform-specific docs repo.